Improve splitting of hplip with new `-drivers` subpackage
Not everyone needs or wants the UI for hplip, but need the scan and printer
drivers and would happily use gnome-control-center to add the printers to
their system.

To facilitate this, we've split the main bulk of the drivers (non GUI bits)
into `hplip-drivers`. Additionally, this means the drivers can now be shipped
as part of an ISO without a massive dependency chain instigated by the use
of `python3-qt5`, which inflates to the ISO to an unnecessary size.

Maniphest Tasks: T4258

Signed-off-by: Ikey Doherty <ikey@solus-project.com>